SCB-527 adjust jaxrs structure
