+++ archetypes/business-service-jaxrs/src/main/resources/archetype-resources/src/main/resources/META-INF/spring/provider.bean.xml
@@ -22,6 +22,6 @@
   xsi:schemaLocation="http://www.springframework.org/schema/beans classpath:org/springframework/beans/factory/xml/spring-beans-3.0.xsd
 		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd">
 
-  <!--change to your endpoint package-->
-  <context:component-scan base-package="org.apache.servicecomb.archetypes"/>
+  <!--change to your schema(impl) bean package-->
+  <context:component-scan base-package="${groupId}"/>
 </beans>
