+++ archetypes/business-service-jaxrs/src/main/resources/archetype-resources/src/main/resources/microservice.yaml
